Extract eslint extensions to module
