+++ packages/brookjs-cli/src/commands/TestCommand/Lint/types.tsx
@@ -1,9 +1,9 @@
 import { Maybe } from 'brookjs-types';
 import { ActionType } from 'typesafe-actions';
 import * as t from 'io-ts';
-import * as glob from '../../../glob';
-import * as actions from './actions';
 import { CLIEngine } from 'eslint';
+import * as glob from '../../../glob';
+import * as eslint from '../../../eslint';
 
 export const RC = t.partial({
   dir: t.string
@@ -24,11 +24,12 @@ export type FileLint =
       status: 'errored';
       error: Error;
     };
+
 export type State = {
   cwd: string;
   rc: Maybe<RC>;
   status: 'globbing' | 'linting' | 'completed';
   files: FileLint[];
 };
-export type Action = ActionType<typeof actions & typeof glob.actions>;
+export type Action = ActionType<typeof glob.actions & typeof eslint.actions>;
 export type Args = {};
