+++ packages/brookjs-cli/src/index.ts
@@ -1,7 +1,8 @@
 import * as glob from './glob';
+import * as eslint from './eslint';
 
 export * from './services';
 export * from './cli';
 export { default as create } from './create';
 
-export { glob };
+export { glob, eslint };
